moved Matchup class to own file in attempt to organize package
