+++ espnff/espnff.py
@@ -4,6 +4,7 @@ from .utils import (two_step_dominance,
                     power_points, )
 from .team import Team
 from .settings import Settings
+from .matchup import Matchup
 
 
 class ESPNFFException(Exception):
@@ -133,31 +134,3 @@ class League(object):
                     matchup.away_team = team
 
         return result
-
-
-class Matchup(object):
-    '''Creates Matchup instance'''
-    def __init__(self, data):
-        self.data = data
-        self._fetch_matchup_info()
-
-    def __repr__(self):
-        return 'Matchup(%s, %s)' % (self.home_team, self.away_team, )
-
-    def _fetch_matchup_info(self):
-        '''Fetch info for matchup'''
-        if self.data['teams'][0]['home'] and not self.data['bye']:
-            self.home_team = self.data['teams'][0]['teamId']
-            self.home_score = self.data['teams'][0]['score']
-            self.away_team = self.data['teams'][1]['teamId']
-            self.away_score = self.data['teams'][1]['score']
-        elif self.data['teams'][0]['home'] and not self.data['bye']:
-            self.home_team = self.data['teams'][1]['teamId']
-            self.home_score = self.data['teams'][1]['score']
-            self.away_team = self.data['teams'][0]['teamId']
-            self.away_score = self.data['teams'][0]['score']
-        else:
-            self.home_team = self.data['teams'][0]['teamId']
-            self.home_score = self.data['teams'][0]['score']
-            self.away_team = None
-            self.away_score = None
